requirements.txt
asgiref==3.7.2
certifi==2023.7.22
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1.4
django-heroku==0.3.1
gunicorn==20.0.4
idna==2.10
numpy==1.25.0
pandas==2.0.3
psycopg2==2.9.7
python-dateutil==2.8.2
pytz==2023.3
requests==2.25.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.16
whitenoise==6.5.0
djangorestframework==3.14.0